changed type of id from uint to uint64

+++ db/models/contact.go
@@ -1,7 +1,7 @@
 package models
 
 type Contact struct {
-	UserID      uint   `gorm:"column:user_id"`
-	ContactID   uint   `gorm:"column:contact_id"`
+	UserID      uint64 `gorm:"column:user_id"`
+	ContactID   uint64 `gorm:"column:contact_id"`
 	ContactName string `gorm:"column:contact_name"`
 }
